+++ NEWS.md
@@ -1,3 +1,17 @@
+Version 1.0.2 (2024.01.23)
+--------------------------
+
+* Added bar charts of frequencies of executed `Command`-s, `Construction` instructions, and the ratio of main/alternative `Branch` choices over interval of time preceding current instant. This window is active while `run`ning.
+
+![freqs scast gif](https://raw.githubusercontent.com/aelhometta/visuals/main/aelhometta_freqs.gif)
+
+Since the spectacle of running Ælhometta takes most of the time, let it be a little less monotonous.
+
+* Added total counts of `Construction` instructions during `Construct` executions, of main/alternative `Branch` choices and of `Space`-s during ticks, to statistics displayed by `@ stat tick`, which replaced `@ stat comm`.
+
+* Fixed encoding of `IntegerToIntegerChannel`, `IntegerToIntegerIndex` and `ShiftDown`, `ShiftUp` commands accordingly to alphabetical order.
+
+
 Version 1.0.0 (2024.01.16)
 --------------------------